fix: lambda_weights fallback missing dtype — matches float32 regardless of target
The else branch created torch.ones(1000, device=device) without dtype,
defaulting to float32 even when dtype=float16/bfloat16 was passed.
The importance_weights fallback on the next line already included
dtype=dtype correctly — this brings lambda_weights in line.

+++ trainer/derrian_backend/sd_scripts/library/edm2_loss.py
@@ -69,7 +69,7 @@ class AdaptiveLossWeightMLP(nn.Module):
         self.a_bar_std = self.alphas_cumprod.std()
         self.logvar_fourier = FourierFeatureExtractor(logvar_channels, dtype=dtype)
         self.logvar_linear = NormalizedLinearLayer(logvar_channels, 1, kernel=[], dtype=dtype) # kernel = []? (not in code given, added matching edm2)
-        self.lambda_weights = lambda_weights.to(device=device, dtype=dtype) if lambda_weights is not None else torch.ones(1000, device=device)
+        self.lambda_weights = lambda_weights.to(device=device, dtype=dtype) if lambda_weights is not None else torch.ones(1000, device=device, dtype=dtype)
         self.noise_scheduler = noise_scheduler
         self.dtype=dtype
